+++ sdk/resourcemanager/appconfiguration/armappconfiguration/client_factory.go
@@ -66,9 +66,3 @@ func (c *ClientFactory) NewPrivateLinkResourcesClient() *PrivateLinkResourcesCli
 	subClient, _ := NewPrivateLinkResourcesClient(c.subscriptionID, c.credential, c.options)
 	return subClient
 }
-
-// NewReplicasClient creates a new instance of ReplicasClient.
-func (c *ClientFactory) NewReplicasClient() *ReplicasClient {
-	subClient, _ := NewReplicasClient(c.subscriptionID, c.credential, c.options)
-	return subClient
-}