+++ sdk/resourcemanager/appconfiguration/armappconfiguration/constants.go
@@ -10,7 +10,7 @@ package armappconfiguration
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/appconfiguration/armappconfiguration"
-	moduleVersion = "v2.1.0"
+	moduleVersion = "v3.0.0"
 )
 
 // ActionsRequired - Any action that is required beyond basic workflow (approve/ reject/ disconnect)
@@ -158,27 +158,5 @@ func PossiblePublicNetworkAccessValues() []PublicNetworkAccess {
 	return []PublicNetworkAccess{
 		PublicNetworkAccessDisabled,
 		PublicNetworkAccessEnabled,
-	}
-}
-
-// ReplicaProvisioningState - The provisioning state of the replica.
-type ReplicaProvisioningState string
-
-const (
-	ReplicaProvisioningStateCanceled  ReplicaProvisioningState = "Canceled"
-	ReplicaProvisioningStateCreating  ReplicaProvisioningState = "Creating"
-	ReplicaProvisioningStateDeleting  ReplicaProvisioningState = "Deleting"
-	ReplicaProvisioningStateFailed    ReplicaProvisioningState = "Failed"
-	ReplicaProvisioningStateSucceeded ReplicaProvisioningState = "Succeeded"
-)
-
-// PossibleReplicaProvisioningStateValues returns the possible values for the ReplicaProvisioningState const type.
-func PossibleReplicaProvisioningStateValues() []ReplicaProvisioningState {
-	return []ReplicaProvisioningState{
-		ReplicaProvisioningStateCanceled,
-		ReplicaProvisioningStateCreating,
-		ReplicaProvisioningStateDeleting,
-		ReplicaProvisioningStateFailed,
-		ReplicaProvisioningStateSucceeded,
 	}
 }