+++ sdk/resourcemanager/appconfiguration/armappconfiguration/fake/server_factory.go
@@ -24,7 +24,6 @@ type ServerFactory struct {
 	OperationsServer                 OperationsServer
 	PrivateEndpointConnectionsServer PrivateEndpointConnectionsServer
 	PrivateLinkResourcesServer       PrivateLinkResourcesServer
-	ReplicasServer                   ReplicasServer
 }
 
 // NewServerFactoryTransport creates a new instance of ServerFactoryTransport with the provided implementation.
@@ -46,7 +45,6 @@ type ServerFactoryTransport struct {
 	trOperationsServer                 *OperationsServerTransport
 	trPrivateEndpointConnectionsServer *PrivateEndpointConnectionsServerTransport
 	trPrivateLinkResourcesServer       *PrivateLinkResourcesServerTransport
-	trReplicasServer                   *ReplicasServerTransport
 }
 
 // Do implements the policy.Transporter interface for ServerFactoryTransport.
@@ -83,9 +81,6 @@ func (s *ServerFactoryTransport) Do(req *http.Request) (*http.Response, error) {
 			return NewPrivateLinkResourcesServerTransport(&s.srv.PrivateLinkResourcesServer)
 		})
 		resp, err = s.trPrivateLinkResourcesServer.Do(req)
-	case "ReplicasClient":
-		initServer(s, &s.trReplicasServer, func() *ReplicasServerTransport { return NewReplicasServerTransport(&s.srv.ReplicasServer) })
-		resp, err = s.trReplicasServer.Do(req)
 	default:
 		err = fmt.Errorf("unhandled client %s", client)
 	}